+++ src/Entity/Track/WayPoint.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Entity\Track;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Entity
+ */
+class WayPoint
+{
+    use PointTrait;
+
+    /**
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="UUID")
+     * @ORM\Column(name="id", type="guid")
+     */
+    private $id;
+
+    public function __construct(float $lat, float $lng)
+    {
+        $this->lat = $lat;
+        $this->lng = $lng;
+    }
+
+    public function getId(): ?string
+    {
+        return $this->id;
+    }
+}
